+++ students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/Task1.java
@@ -1,16 +1,15 @@
-package ru.mirea.practice.s21K0689;
+package ru.mirea.practice;
 
 public class Task1 {
-    public static void main(String[] args)
-    {
+    public static void main(String[] args) {
         int [] a = {1,23,4,235,346,123};
-        double sum =0;
+        double sum = 0;
         double average;
         for (int j : a) {
             System.out.println(j);
             sum += j;
         }
-        average = sum/a.length;
+        average = sum / a.length;
         System.out.println("sum = " + sum);
         System.out.println("average = " + average);
     }
