+++ students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/Task5.java
@@ -1,0 +1,13 @@
+package ru.mirea.practice;
+public class Task5 {
+    public static void main(String[] args) {
+        for (int i = 1; i <= 10; i++) {
+            if (i == 1) {
+                System.out.println(i);
+            }
+            else {
+                System.out.println("1/" + i);
+            }
+        }
+    }
+}
