--- students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/s21K0689/Task3.java
@@ -1,11 +1,0 @@
-package ru.mirea.practice.s21K0689;
-
-public class Task3 {
-    public static void main (String [] args)
-    {
-        for (String str:args)
-        {
-            System.out.println("аргумент = " +str);
-        }
-    }
-}
