--- students/21K0689/21K0689-p01/src/main/java/ru/mirea/practice/s21K0689/Task4.java
@@ -1,15 +1,0 @@
-package ru.mirea.practice.s21K0689;
-
-public class Task4 {
-    public static void main(String [] args)
-    {
-        for (int i = 1; i<=10; i++)
-        {
-            if (i==1)
-            {
-                System.out.println(i);
-            }
-            else System.out.println("1/" + i);
-        }
-    }
-}
